Move tests into its own file
